Enable "semantic hashing" of the Pipfile
Instead of hashing the actual bytes of the ``Pipfile`` instance, parse
it first and then serialize that parsed instance into a canonical
format before hashing. This will make it so things like adding comments,
whitespace, re-arranging table keys, and other changes that don't alter
the *semantics* of the ``Pipfile`` do not invalidate a ``Pipfile.lock``.